# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  internal_api.CreateRoomRequest:
    properties:
      capacity:
        type: integer
      name:
        type: string
      room_type:
        type: string
    type: object
...
  internal_api.MagicLinkRequest:
    properties:
      email:
        type: string
    type: object
  internal_api.MagicLinkResponse:
    properties:
      code:
        type: string
      errors:
        additionalProperties: true
        type: object
      message:
        type: string
    type: object
...
  internal_api.UpdateRoomRequest:
    properties:
      capacity:
        type: integer
      id:
        type: string
      name:
        type: string
      room_type:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: Platform team
  description: Backend REST API for Somotracker educational dashboard
  license:
    name: MIT
  termsOfService: http://somotracker.local/terms
  title: Somotracker API
  version: 1.0.0
paths:
...
  /api/auth/magic-link/send:
    post:
      consumes:
      - application/json
      parameters:
      - description: User email
        in: formData
        name: email
        type: string
      - description: JSON body
        in: body
        name: body
        schema:
          $ref: '#/definitions/internal_api.MagicLinkRequest'
      - description: Organization ID or slug
        in: query
        name: org_id
        type: string
      - description: Magic link request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.MagicLinkRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.MagicLinkResponse'
      summary: Send magic-link email
      tags:
      - Auth
...
  /school/rooms:
    get:
      description: List all rooms for the active school with pagination
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: List rooms
      tags:
      - Rooms
    patch:
      consumes:
      - application/json
      description: Update room details for the active school
      parameters:
      - description: Room update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.UpdateRoomRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: Update room
      tags:
      - Rooms
    post:
      consumes:
      - application/json
      description: Create a new room for the active school
      parameters:
      - description: Room details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.CreateRoomRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
      summary: Create room
      tags:
      - Rooms
  /school/rooms/{id}:
    delete:
      description: Delete a room for the active school
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: Delete room
      tags:
      - Rooms
    get:
      description: Get a room by ID for the active school
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: Get room
      tags:
      - Rooms
...
swagger: "2.0"